updated delete_goal func to account for 404, corrected wave 5 test/ passed wave 5

+++ app/routes.py
@@ -220,14 +220,17 @@ def create_goal():
 # update existing goal
 @goals_bp.route("/<goal_id>", methods=["PUT"])
 def update_goal(goal_id):
-    goal = Goal.query.get_or_404(goal_id)
+    goal = Goal.query.get(goal_id)
+
+    if goal is None:
+        return jsonify({"details": f"Goal {goal_id} was not found."}), 404
+
     request_data = request.get_json()
-    if "title" in request_data:
-        goal.title = request_data["title"]
-        
-        db.session.commit()
+    goal.title = request_data.get("title", goal.title)
+    db.session.commit()
 
     return jsonify({"goal": goal.to_dict()})
+
 
 # get specific goal/ account for 404
 @goals_bp.route("/<goal_id>", methods=["GET"])
@@ -239,15 +242,18 @@ def get_goal(goal_id):
 
     return jsonify({"goal": goal.to_dict()})
 
+
 # delete a goal
 @goals_bp.route("/<goal_id>", methods=["DELETE"])
 def delete_goal(goal_id):
-    goal = Goal.query.get_or_404(goal_id)
+    goal = Goal.query.get(goal_id)
+
+    if goal is None:
+        return jsonify({"details": f"Goal {goal_id} was not found."}), 404
+
     db.session.delete(goal)
     db.session.commit()
 
     return jsonify({
         "details": f'Goal {goal.id} "{goal.title}" successfully deleted'
     }), 200
-    
-    #
